requirements.txt
python-dotenv==1.1.0
requests==2.32.3
pytz==2025.2
rich==14.0.0
GitPython==3.1.44
openai==1.86.0
typer==0.16.0
